+++ im_flutter_sdk/lib/src/internal/chat_method_keys.dart
@@ -424,6 +424,10 @@ class ChatMethodKeys {
 
   static const String bindDeviceToken = 'bindDeviceToken';
 
+  // 4.25.0
+  static const String bindPushKitToken = 'bindPushKitToken';
+  static const String unbindPushKitToken = 'unbindPushKitToken';
+
   // 4.10
   static const String onOfflineMessageSyncStart = 'onOfflineMessageSyncStart';
   static const String onOfflineMessageSyncFinish = 'onOfflineMessageSyncFinish';