+++ im_flutter_sdk_ios/ios/im_flutter_sdk_ios/Sources/im_flutter_sdk_ios/include/im_flutter_sdk_ios/MethodKeys.h
@@ -386,6 +386,10 @@ static NSString *const ChatUpdateRegradeMessagesSetting =
     @"updateRegradeMessagesSetting";
 static NSString *const bindDeviceToken = @"bindDeviceToken";
 
+// 4.25.0
+static NSString *const bindPushKitToken = @"bindPushKitToken";
+static NSString *const unbindPushKitToken = @"unbindPushKitToken";
+
 
 // 4.10
 static NSString *const onOfflineMessageSyncStart = @"onOfflineMessageSyncStart";